diploma_pj/requirements-dev.txt
-r requirements.txt
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
